Add threads and overwrite options

+++ clustering_experiments/commands.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 usage="USAGE:\n
-                $0 genomes.csv path/to/RefPkgs/ [threads=8]"
+                $0 -g [genomes.csv] -r [path/to/RefPkgs/] [-t num_threads=8] [-o overwrite='y'|'n']"
 
 if [ $# -eq 0 ]; then
 	echo "ERROR: No arguments provided."
@@ -10,26 +10,48 @@ if [ $# -eq 0 ]; then
 elif [ $1 == "-h" ]; then
 	echo -e $usage
 	exit
-elif [ $# -lt 2 ]; then
+elif [ $# -lt 4 ]; then
 	echo "ERROR: Insufficient number of arguments provided."
-	echo $usage
+	echo -e $usage
+	exit
 fi
 
 # Set arguments
-genomes_list=$1
-refpkgs_repo_dir=$2
-if [ -z $3 ]; then
+while getopts ":g:r:t:o:" opt; do
+  case $opt in
+  g) genomes_list="$OPTARG"
+    ;;
+  r) refpkgs_repo_dir="$OPTARG"
+    ;;
+  t) n_threads="$OPTARG"
+    ;;
+  o) overwrite="$OPTARG"
+    ;;
+  \?) echo "Invalid option -$OPTARG" >&2
+    ;;
+  esac
+done
+if [ -z $n_threads ]; then
   n_threads=8
-else
-  n_threads=$3
+fi
+if [ -z $overwrite ]; then
+  overwrite="y"
 fi
 
 proteins_fa=proteome.fasta
+log_file=TreeSAPP_clustering_experiments_log.txt
 refpkg_dir=clustering_refpkgs/
 ts_assign_out=proteome_classified
 phylotu_out=phylotu_outputs/
 
-if [ -f $proteins_fa ]; then
+printf "Running with the following configuration:
+        Genomes list: $genomes_list
+        RefPkgs path: $refpkgs_repo_dir
+        Overwrite: $overwrite
+        Threads: $n_threads
+        Log file: $log_file\n"
+
+if [ -f $proteins_fa ] && [ $overwrite == 'y' ]; then
 	rm $proteins_fa
 fi
 if [ ! -d $refpkg_dir ]; then
@@ -40,18 +62,22 @@ if [ ! -d $phylotu_out ]; then
 fi
 
 # Download genomes
-while read line
-do
-	taxid=$( echo "$line" | gawk -F, '{ print $2 }')
-	faa_url=$( echo "$line" | gawk -F, '{ print $4 }')
-	if [ $faa_url == "ProteinsURL" ]; then
-	  continue
-	fi
-	echo "Downloading sequences for $( echo "$line" | gawk -F, '{ print $1 }')"
-	wget -O $taxid.faa.gz $faa_url 1>/dev/null 2>&1
-	gunzip -c $taxid.faa.gz | seqkit replace -p "^" -r "${taxid}." >>$proteins_fa
-	rm $taxid.faa.gz
-done<$genomes_list
+if [ ! -f $proteins_fa ]; then
+  while read line
+  do
+    taxid=$( echo "$line" | gawk -F, '{ print $2 }')
+    faa_url=$( echo "$line" | gawk -F, '{ print $4 }')
+    if [ $faa_url == "ProteinsURL" ]; then
+      continue
+    fi
+    echo "Downloading sequences for $( echo "$line" | gawk -F, '{ print $1 }')"
+    wget -O $taxid.faa.gz $faa_url 1>/dev/null 2>&1
+    gunzip -c $taxid.faa.gz | seqkit replace -p "^" -r "${taxid}." >>$proteins_fa
+    rm $taxid.faa.gz
+  done<$genomes_list
+else
+  echo "Using $proteins_fa from previous run"
+fi
 
 echo "Copying reference packages"
 cp \
@@ -65,11 +91,14 @@ cp \
 for r in $(seq 100 50 600)
 do
 	prefix=length_$r
-	if [ -d $prefix ]; then
-	       rm -r $prefix
-        fi
-	mkdir $prefix
-	mkdir $prefix/$phylotu_out
+	if [ -d $prefix ] && [ $overwrite == 'y' ]; then
+	  rm -r $prefix
+    mkdir $prefix
+    mkdir $prefix/$phylotu_out
+	elif [ ! -d $prefix ]; then
+    mkdir $prefix
+    mkdir $prefix/$phylotu_out
+  fi
 
   queries=$prefix/$proteins_fa
   overlap=$( echo $r | awk '{ print $1/2 }' )
@@ -78,33 +107,62 @@ do
   seqkit sliding --greedy --step $overlap --window $r | \
   seqkit seq  --remove-gaps --min-len 30 >$queries
 
-	# Classify the sequences
-	if [ ! -d $prefix/$ts_assign_out ]; then 
+	if [ ! -d $prefix/$ts_assign_out ]; then
+    echo "Classifying the sequences"
 		treesapp assign \
 		--fastx_input $queries \
 		--output $prefix/$ts_assign_out \
 		--refpkg_dir $refpkg_dir \
 		-m prot --num_procs $n_threads --delete --overwrite
+	else
+	  echo "Using treesapp assign classifications in $prefix/$ts_assign_out"
 	fi
 
-  # Cluster the classified query sequences
+  echo "Clustering the classified query sequences"
 	for f in $refpkg_dir/*pkl
 	do
 	  for rank in "species" "family" "class"
 	  do
+	    # Output directories
+	    potu_rg_out=$prefix/$phylotu_out/$( basename $f | sed 's/_build.pkl//g' )\_phylotus_rg_$rank
+	    potu_dn_psc_out=$prefix/$phylotu_out/$( basename $f | sed 's/_build.pkl//g' )\_phylotus_dn_psc_$rank
+	    potu_dn_aln_out=$prefix/$phylotu_out/$( basename $f | sed 's/_build.pkl//g' )\_phylotus_dn_aln_$rank
+
 	    # Reference-guided based on placement edges
-      treesapp phylotu \
-        --refpkg_path $f \
-        --assign_output $prefix/$ts_assign_out \
-        --tax_rank $rank \
-        -o $prefix/$phylotu_out/$( basename $f | sed 's/_build.pkl//g' )\_phylotus_rg_$rank
+	    if [ -d $potu_rg_out ] && [ $overwrite == 'n' ]; then
+	      echo "Skipping $potu_rg_out"
+	    else
+        treesapp phylotu \
+          --refpkg_path $f \
+          --assign_output $prefix/$ts_assign_out \
+          --tax_rank $rank \
+          -o $potu_rg_out \
+          --mode ref_guided
+      fi
+
       # De novo clusters by recreating the phylogeny
-      treesapp phylotu \
-        --refpkg_path $f \
-        --assign_output $prefix/$ts_assign_out \
-        --tax_rank $rank \
-        -o $prefix/$phylotu_out/$( basename $f | sed 's/_build.pkl//g' )\_phylotus_dn_$rank \
-        --mode de_novo
+      if [ -d $potu_dn_psc_out ] && [ $overwrite == 'n' ]; then
+	      echo "Skipping $potu_dn_psc_out"
+	    else
+        treesapp phylotu \
+          --refpkg_path $f \
+          --assign_output $prefix/$ts_assign_out \
+          --tax_rank $rank \
+          -o $potu_dn_psc_out \
+          --mode de_novo --pre_cluster psc
+      fi
+
+      # De novo clusters by recreating the phylogeny
+      if [ -d $potu_dn_aln_out ] && [ $overwrite == 'n' ]; then
+	      echo "Skipping $potu_dn_aln_out"
+	    else
+        treesapp phylotu \
+          --refpkg_path $f \
+          --assign_output $prefix/$ts_assign_out \
+          --tax_rank $rank \
+          -o $potu_dn_aln_out \
+          --mode de_novo --pre_cluster align
+      fi
     done
   done
 done
